Refactor trends presenter into modular helpers
Tests: npm test -- tests/ui/views/browser/cardsPresenter.test.js
